+++ .github/workflows/issue-trigger.yml
@@ -13,7 +13,7 @@ jobs:
         uses: Aveline-art/bookish-umbrella@main
         with:
           myToken: ${{ secrets.GITHUB_TOKEN }}
-          labels: dependency
+          label-string: dependency
           message: 'This is an issue message'
       # Use the output from the `hello` step
       - name: Get the output time
